py/services/users/pyproject.toml
[tool.poetry]
name = "users"
version = "0.1.0"
description = ""
authors = ["Aleksey Dalekin <adalekin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.9"
connexion-buzz = "0.2.1"
connexion = "2.9.0"
environs = "9.3.2"
flask-marshmallow = "0.14.0"
Flask-Migrate = "3.1.0"
Flask-SQLAlchemy = "2.5.1"
gunicorn = "20.0.4"
marshmallow-sqlalchemy = "0.26.1"
marshmallow = "3.13.0"
psycopg2 = "2.9.1"
sentry-sdk = {version = "0.7.10", extras = ["flask"]}
SQLAlchemy = "1.4.22"
talepy = "0.8.0"
flask-kafka = {path = "../../lib/flask-kafka"}
schemas = {path = "../../lib/schemas"}

[tool.poetry.dev-dependencies]
coverage = "4.5.1"
pytest-cov = "2.7.1"
pytest-env = "0.6.2"
pytest-freezegun = "0.3.0.post1"
pytest-mock = "1.10.1"
pytest-toolbelt = "0.1.0"
pytest = "3.9.1"
requests-mock = "1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
